[fuchsia] Make the component controller compile.
